feat(matcher): E1.3a — move the matcher's pure leaf deps into exact_matching
Move the 4 pure-logic leaf modules the matcher depends on (zero trials.models /
ORM refs) into the package + shims: enums, utils, patient_info.genetic_mutations,
patient_info.patient_info_flipi_score. The matcher's imports of these repoint to
the package; old paths are import-* shims so all other importers (7 utils, 4
enums, 2+2) stay unchanged.

First slice of E1.3 (make exact_matching CB-self-contained). The matcher's two
remaining trials deps — configs and patient_info_attributes — drag larger
subtrees (convertors/, mappers) and move in E1.3b/c; the TrialMatchingQuerySet +
geo abstract-model cluster is E1.3d.

Full EXACT suite green (1073 passed, 5 skipped).

Part of EPIC cancerbot#4601 (E1 #272, D1 resolved). Proposal — cross-team.
