+++ trials/services/patient_info/patient_info_flipi_score.py
@@ -1,8 +1,2 @@
-class PatientInfoFlipyScore:
-    @staticmethod
-    def scope_by_options(flipi_score_options):
-        avail_options = ['age', 'stage', 'hemoglobin', 'nodalAreas', 'ldh']
-        items = str(flipi_score_options).split(',')
-        items = [x.strip() for x in items]
-        res = len(set(items).intersection(avail_options))
-        return res if res > 0 else None
+"""Back-compat shim — moved to exact_matching.patient_info.patient_info_flipi_score (E1.3a)."""
+from exact_matching.patient_info.patient_info_flipi_score import *  # noqa: F401,F403
